Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/error-type-2-expected.txt

CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '%'

FAIL parse error has higher priority than instantiation error assert_equals: expected 4 but got 3
